+++ src/core/index.ts
@@ -1,0 +1,4 @@
+import DynamicEngine from './DynamicEngine';
+import ViewRender from './renderer/ViewRender';
+import FormRender from './renderer/FormRender';
+export { DynamicEngine, ViewRender, FormRender };
